+++ scripts/globals/utils.js
@@ -1,6 +1,8 @@
 import angular from 'angular';
 import moment from 'moment';
 
+export { moment };
+
 export const mergeObjects = angular.merge;
 
 export const leadZero = function (num) {
@@ -54,9 +56,9 @@ export const playSound = function (sound) {
    audio.play();
 };
 
-export const timeAgo = function (time) {
+export const timeAgo = function (time, withoutSuffix = false) {
    const momentInTime = moment(new Date(time));
-   return momentInTime.fromNow();
+   return momentInTime.fromNow(withoutSuffix);
 };
 
 export const debounce = function (func, wait, immediate) {
@@ -79,9 +81,9 @@ export const debounce = function (func, wait, immediate) {
    };
 };
 
-export const toAbsoluteServerURL = function (path) {
+export const toAbsoluteServerURL = function (path, serverUrlOverride = false) {
    const startsWithProtocol = path.indexOf('http') === 0;
-   const url = startsWithProtocol ? path : window.CONFIG.serverUrl + '/' + path;
+   const url = startsWithProtocol ? path : (serverUrlOverride || window.SERVER_URL_OVERRIDE || window.CONFIG.serverUrl) + '/' + path;
    // Replace extra forward slashes but not in protocol.
    return url.replace(/([^:])\/+/g, '$1/');
 };