+++ scripts/init.js
@@ -28,11 +28,11 @@ window.initApp = function () {
       }
 
       ApiProvider.setInitOptions({
-         wsUrl: window.CONFIG.wsUrl,
-         authToken: window.CONFIG.authToken,
+         wsUrl: window.WS_URL_OVERRIDE || window.CONFIG.wsUrl,
+         authToken: window.AUTH_TOKEN_OVERRIDE || window.CONFIG.authToken,
       });
 
-      tmhDynamicLocaleProvider.localeLocationPattern('./locales/angular-locale_{{locale}}.js');
+      tmhDynamicLocaleProvider.localeLocationPattern('./locales/{{locale}}.js');
 
       const clock24 = window.CONFIG.timeFormat === 24;