+++ types/globals.d.ts
@@ -1,0 +1,6 @@
+interface Window {
+    AUTH_TOKEN_OVERRIDE: string | null
+    REST_URL_OVERRIDE: string | null
+    SERVER_URL_OVERRIDE: string | null
+    WS_URL_OVERRIDE: string | null
+}